etl_pipeline/requirements.txt
pandas

python-dotenv

azure-storage-blob

sqlalchemy
psycopg2-binary

prefect

matplotlib
seaborn
